Set DooFile::delete() to check folder exists before deleting to stop error being triggered Set DooFile::create() to chmod created file with defined setting
git-svn-id: http://doophp.googlecode.com/svn/trunk@423 eaaf3d48-7251-11de-8585-3f51e31374b7

+++ dooframework/helper/DooFile.php
@@ -56,26 +56,28 @@ class DooFile {
      */
 	public function delete($path, $deleteSelf=true){
 
-        //delete all sub folder/files under, then delete the folder itself
-        if(is_dir($path)){
-            if($path[strlen($path)-1] != '/' && $path[strlen($path)-1] != '\\' ){
-                $path .= DIRECTORY_SEPARATOR;
-                $path = str_replace('\\', '/', $path);
-            }
-            if($total = $this->purgeContent($path)){
-                if($deleteSelf)
-                    if($t = rmdir($path))
-                        return $total + $t;
-                return $total;
-            }
-            else if($deleteSelf){
-                return rmdir($path);
-            }
-            return false;
-        }
-        else{
-            return unlink($path);
-        }
+		if (file_exists($path)) {
+			//delete all sub folder/files under, then delete the folder itself
+			if(is_dir($path)){
+				if($path[strlen($path)-1] != '/' && $path[strlen($path)-1] != '\\' ){
+					$path .= DIRECTORY_SEPARATOR;
+					$path = str_replace('\\', '/', $path);
+				}
+				if($total = $this->purgeContent($path)){
+					if($deleteSelf)
+						if($t = rmdir($path))
+							return $total + $t;
+					return $total;
+				}
+				else if($deleteSelf){
+					return rmdir($path);
+				}
+				return false;
+			}
+			else{
+				return unlink($path);
+			}
+		}
     }
 
 
@@ -108,6 +110,7 @@ class DooFile {
             $fp = fopen($filename, $writeFileMode);
             $rs = fwrite($fp, $content);
             fclose($fp);
+			chmod($filename, $this->chmod);
             return ($rs>0);
 		}else{
 			if (!file_exists($path)) {
